Chore: merge all functions using autorun x useSubscription pattern (#26886)
