+++ apps/meteor/client/lib/createSubscription.ts
@@ -1,0 +1,7 @@
+import { createReactiveSubscriptionFactory } from './createReactiveSubscriptionFactory';
+
+export const createSubscription = function <T>(
+	getValue: () => T,
+): [subscribe: (onStoreChange: () => void) => () => void, getSnapshot: () => T] {
+	return createReactiveSubscriptionFactory(getValue)();
+};
