+++ apps/meteor/client/providers/SessionProvider.tsx
@@ -2,7 +2,7 @@ import { SessionContext } from '@rocket.chat/ui-contexts';
 import { Session } from 'meteor/session';
 import React, { FC } from 'react';
 
-import { createReactiveSubscriptionFactory } from './createReactiveSubscriptionFactory';
+import { createReactiveSubscriptionFactory } from '../lib/createReactiveSubscriptionFactory';
 
 const contextValue = {
 	query: createReactiveSubscriptionFactory<unknown>((name) => Session.get(name)),
