+++ apps/meteor/client/providers/SettingsProvider.tsx
@@ -2,9 +2,9 @@ import { SettingsContext, SettingsContextValue, useAtLeastOnePermission, useMeth
 import { Tracker } from 'meteor/tracker';
 import React, { useCallback, useEffect, useMemo, useState, FunctionComponent } from 'react';
 
+import { createReactiveSubscriptionFactory } from '../lib/createReactiveSubscriptionFactory';
 import { PrivateSettingsCachedCollection } from '../lib/settings/PrivateSettingsCachedCollection';
 import { PublicSettingsCachedCollection } from '../lib/settings/PublicSettingsCachedCollection';
-import { createReactiveSubscriptionFactory } from './createReactiveSubscriptionFactory';
 
 type SettingsProviderProps = {
 	readonly privileged?: boolean;
