+++ apps/meteor/client/providers/UserProvider.tsx
@@ -7,7 +7,7 @@ import { Subscriptions, Rooms } from '../../app/models/client';
 import { getUserPreference } from '../../app/utils/client';
 import { callbacks } from '../../lib/callbacks';
 import { useReactiveValue } from '../hooks/useReactiveValue';
-import { createReactiveSubscriptionFactory } from './createReactiveSubscriptionFactory';
+import { createReactiveSubscriptionFactory } from '../lib/createReactiveSubscriptionFactory';
 
 const getUserId = (): string | null => Meteor.userId();
